requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
black==24.3.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
docker==7.0.0
execnet==2.0.2
Faker==37.0.0
filelock==3.13.1
identify==2.5.33
idna==3.7
iniconfig==2.0.0
marshmallow==3.20.1
marshmallow-dataclass==8.6.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.2
pyright==1.1.352
pytest==8.0.2
pytest-instafail==0.5.0
pytest-rerunfailures==13.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
python-dotenv==1.0.1
pytest-dependency==0.6.0
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.21
scikit-learn~=1.6.1
setuptools==70.0.0
tenacity==8.2.3
typeguard==4.1.5
typing-inspect==0.9.0
typing_extensions>=4.10
urllib3==2.2.2
virtualenv==20.25.0
Jinja2~=3.1.5
psutil~=7.0.0
pytest-shard==0.1.2
learn~=1.0.0
pandas~=2.3.0
nltk~=3.9.1